requirements.txt
aiofiles==23.2.1
aiogram==3.12.0
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
frozenlist==1.4.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
mutagen==1.47.0
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
yt-dlp==2024.8.6
